requirements.txt
requests
2captcha-python
faker
python-dotenv
loguru
colorama
tabulate